Market Hours

api_getMarketHoursAsJSON

TDAmeritradeAPI.api_getMarketHoursAsJSON — Function

api_getMarketHoursAsJSON(market::String, apiKeys::apiKeys; kw...)::ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_market_hours_for_single_market endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

market::String: The market to get hours for. Valid values are:

"EQUITY"

"OPTION"

"FUTURE"

"BOND"

"FOREX"

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

marketDate::Date: Date to fetch the market hours for. Default to today(), unless today is a weekend, then the next Monday.

Example

api_getMarketHoursAsJSON("EQUITY", apiKey)
Result{String, String}(Ok("{"equity":{"EQ":{"date":"2023-02-17","marketType":"EQUITY","exchange":"NULL","category":"NULL","product":"EQ"
[...]

api_getMarketHoursAsJSON(markets::Vector{String}, apiKeys::apiKeys; kw...)::ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_market_hours_for_multiple_markets endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

markets::Vector{String}: A Vector of markets to get hours for. Valid values are:

"EQUITY"

"OPTION"

"FUTURE"

"BOND"

"FOREX"

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

marketDate::Date: Date to fetch the market hours for. Default to today(), unless today is a weekend, then the next Monday.

Example

api_getMarketHoursAsJSON(["EQUITY", "OPTION"], apiKey)
Result{String, String}(Ok("{"option":{"EQO":{"date":"2023-02-17","marketType":"OPTION","exchange":"NULL","category":"NULL","product":"EQO"
[...]

api_getMarketHoursAsDataFrame

TDAmeritradeAPI.api_getMarketHoursAsDataFrame — Function

api_getMarketHoursAsDataFrame(market::String, apiKeys::apiKeys; kw...)::ErrorTypes.Option{DataFrame}

Make the TDAmeritradeAPI call to the get_market_hours_for_single_market endpoint, and return a DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

See api_getMarketHoursAsJSON(::String, ::apiKeys).

Keywords

See api_getMarketHoursAsJSON(::String, ::apiKeys).

Example

api_getMarketHoursAsDataFrame("EQUITY", apiKey)
some(1x15 DataFrame
 Row | date        marketType  exchange  category  product  productName  isOpen  preMarket_openDateTime ...
     | Date        String      String    String    String   String       Bool    Union    
 ----------------------------------------------------------------------------------------------------------
   1 | 2023-02-17  EQUITY      NULL      NULL      EQ       equity         true  2023-02-17T07:00:00    ...
[...]

api_getMarketHoursAsDataFrame(markets::Vector{String}, apiKeys::apiKeys; kw...)::ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_market_hours_for_multiple_markets endpoint, and return a DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

See api_getMarketHoursAsJSON(::Vector{String}, ::apiKeys).

Keywords

See api_getMarketHoursAsJSON(::Vector{String}, ::apiKeys).

Example

api_getMarketHoursAsDataFrame(["EQUITY", "OPTION"], apiKey)
some(3x15 DataFrame
 Row | date        marketType  exchange  category  product  productName    isOpen  preMarket_openDateTime ...
     | Date        String      String    String    String   String         Bool    Union
 ------------------------------------------------------------------------------------------------------------
   1 | 2023-02-17  EQUITY      NULL      NULL      EQ       equity           true  2023-02-17T07:00:00    ...
   2 | 2023-02-17  OPTION      NULL      NULL      IND      index option     true
[...]

parseMarketHoursJSONToDataFrame

TDAmeritradeAPI.parseMarketHoursJSONToDataFrame — Function

parseMarketHoursJSONToDataFrame(json_string::String)::ErrorTypes.Option{DataFrame}

Convert the JSON string returned by a TDAmeritradeAPI get_market_hours... API to a DataFrame.

Nested JSON objects will be flattened into columns in the output DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

parseMarketHoursJSONToDataFrame(j)
some(3x15 DataFrame
 Row | date        marketType  exchange  category  product  productName    isOpen  preMarket_openDateTime ...
     | Date        String      String    String    String   String         Bool    Union
 ------------------------------------------------------------------------------------------------------------
   1 | 2023-02-17  EQUITY      NULL      NULL      EQ       equity           true  2023-02-17T07:00:00    ...
   2 | 2023-02-17  OPTION      NULL      NULL      IND      index option     true
[...]

marketHoursToMarketTypesStruct

TDAmeritradeAPI.marketHoursToMarketTypesStruct — Function

marketHoursToMarketTypesStruct(json_string::String)::ErrorTypes.Option{MarketTypes}

Convert the JSON string returned by a TDAmeritradeAPI get_market_hours... API calls to a MarketTypes struct.

This is largely an internal function to allow later conversions to DataFrame with proper type conversions.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

marketHoursToMarketTypesStruct(j)
some(TDAmeritradeAPI.MarketTypes(Dict{String, TDAmeritradeAPI.Market}("EQ" => TDAmeritradeAPI.Market(Date("2023-02-17"), "EQUITY"
[...]  

marketHoursToJSON

TDAmeritradeAPI.marketHoursToJSON — Function

marketHoursToJSON(m::MarketTypes)::ErrorTypes.Option{String}

Convert a MarketTypes struct m to a JSON object.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

The returned JSON will not be in the same format as the initial return from the TDAmeritradeAPI call.

Example

marketHoursToJSON(s)
some("{"equity":{"EQ":{"date":"2023-02-17","marketType":"EQUITY","exchange":"NULL","category":"NULL"
[...]