TDAmeritradeAPI.api_getOptionChainAsJSON — Function

api_getOptionChainAsJSON(symbol::String, apiKeys::TDAmeritradeAPI.apiKeys; contractType::String = "ALL", strikeCount::Int64 = 25, includeQuotes::Bool = false, strategy::String = "SINGLE", interval::Int64 = 10, strike::Union{Float64, Nothing} = nothing, range::String = "ALL", fromDate::Date = today(), toDate::Union{Date, Nothing} = nothing, expMonth::String = "ALL", optionType::String = "ALL" )::ErrorTypes.Option{DataFrame}

Make the TDAmeritradeAPI call to the getoptionchain endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: the underlying stock symbol to fetch the Option Chain for.

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

contractType::String = "ALL": Type of contracts to return in the chain. Can be CALL, PUT, or ALL.

strikeCount::Int64 = 25: The number of strikes to return above and below the at-the-money price.

includeQuotes::Bool = false: Include quotes for options in the option chain. Can be TRUE or FALSE.

strategy::String = "SINGLE": Passing a value returns a Strategy Chain. Possible values are SINGLE, COVERED, VERTICAL, CALENDAR, STRANGLE, STRADDLE, BUTTERFLY, CONDOR, DIAGONAL, COLLAR, or ROLL

interval::Int64 = 10: Strike interval for spread strategy chains

strike::Union{Float64, Nothing} = nothing: Provide a strike price to return options only at that strike price.

range::String = "ALL": Returns options for the given range. Possible values are: ITM: In-the-money NTM: Near-the-money OTM: Out-of-the-money SAK: Strikes Above Market SBK: Strikes Below Market SNK: Strikes Near Market ALL: All Strikes

fromDate::Date = today(): Only return expirations after this date. For strategies, expiration refers to the nearest term expiration in the strategy. Valid ISO-8601 formats are: yyyy-MM-dd and yyyy-MM-dd'T'HH:mm:ssz.'

toDate::Union{Date, Nothing} = nothing: Only return expirations before this date. For strategies, expiration refers to the nearest term expiration in the strategy. Valid ISO-8601 formats are: yyyy-MM-dd and yyyy-MM-dd'T'HH:mm:ssz.'

expMonth::String = "ALL": Return only options expiring in the specified month. Month is given in the three character format. Example: JAN

optionType::String = "ALL": Type of contracts to return. Possible values are: S: Standard contracts NS: Non-standard contracts ALL: All contracts

TDAmeritradeAPI.api_getOptionChainAsDataFrame — Function

api_getOptionChainAsDataFrame(symbol::String, apiKeys::TDAmeritradeAPI.apiKeys; contractType::String = "ALL", strikeCount::Int64 = 25, includeQuotes::Bool = false, strategy::String = "SINGLE", interval::Int64 = 10, strike::Union{Float64, Nothing} = nothing, range::String = "ALL", fromDate::Date = today(), toDate::Union{Date, Nothing} = nothing, expMonth::String = "ALL", optionType::String = "ALL" )::ErrorTypes.Option{DataFrame}

Make the TDAmeritradeAPI call to the getoptionchain endpoint, and return a DataFrame

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

See 'api_getOptionChainAsJSON'.

TDAmeritradeAPI.parseOptionChainJSONToDataFrame — Function

parseOptionChainJSONToDataFrame(json_string::String)::ErrorTypes.Option{DataFrame}

Convert the JSON string returned by the TDAmeritradeAPI getoptionchain API call to a DataFrame. The put and call maps will be appended into a single DataFrame, with the PUT rows coming first. Nested JSON objects will be flattened into columns in the output DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

parseOptionChainJSONToDataFrame(js)
some(596x52 DataFrame
 Row | Status  Underlying UnderlyingPrice putCall symbol description exchangeName bid ask
[...]

TDAmeritradeAPI.optionChainToOptionChainStruct — Function

optionChainToOptionChainStruct(json_string::String)::ErrorTypes.Option{OptionChain}

Convert the JSON string returned by the TDAmeritradeAPI getoptionchain API call to an OptionChain struct. This is largely an internal function to allow later conversions to DataFrame with proper type conversions.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

optionChainToOptionChainStruct(js)
some(TDAmeritradeAPI.OptionChain("NET", "SUCCESS", nothing, "SINGLE", 0.0, true, false, 0.1,
[...]

TDAmeritradeAPI.optionChainToJSON — Function

optionChainToJSON(oc::OptionChain)::ErrorTypes.Option{String}

Convert an OptionChain struct oc to a JSON object.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

The returned JSON will not be in the same format as the initial return from the TDAmeritradeAPI call.

Example

TDAmeritradeAPI.optionChainToJSON(oc)
some("{"symbol":"NET","status":"SUCCESS","underlying":null,"strategy":"SINGLE" 
[...]