Price History

api_getPriceHistoryAsJSON

TDAmeritradeAPI.api_getPriceHistoryAsJSON — Function

api_getPriceHistoryAsJSON(symbol::String, keys::apiKeys; kw...)

Make the TDAmeritradeAPI call to the get_price_history endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: The ticker symbol to fetch price history for.

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

periodType::String: The type of period to show. (default in bold):

Valid values are day, month, year, or ytd (year to date)

numPeriods::Int64: The number of periods to show.

Valid periods by periodType (defaults in bold):

day: 1, 2, 3, 4, 5, 10

month: 1, 2, 3, 6

year: 1, 2, 3, 5, 10, 15, 20

ytd: 1

frequencyType::String: The type of frequency with which a new candle is formed.

Valid frequencyTypes by periodType (defaults in bold):

day: minute

month: daily, weekly

year: daily, weekly, monthly

ytd: daily, weekly

frequency::Int64: The number of the frequencyType to be included in each candle.

Valid frequencies by frequencyType (defaults in bold):

minute: 1, 5, 10, 15, 30

daily: 1

weekly: 1

monthly: 1

endDate::DateTime: End date of data to fetch. If startDate and endDate are provided, period should not be provided.

startDate::Union{DateTime, Nothing}: Start date of data to fetch.

needExtendedHoursData: true to return extended hours data, false for regular market hours only. Default is true.

Example

api_getPriceHistoryAsJSON("NET", apiKey)
Result{String, String}(Ok("{"candles":[{"open":60.8,"high":60.8,"low":60.5,"close":60.5,"volume":5000,"datetime":1675425600000},
                           {"open":61.0,"high":61.0,"low":61.0,"close":61.0,"volume":200,"datetime":1675426080000},
                           {"open":61.75,"high":61.8,"low":61.75,"close":61.8,"volume":2347,"datetime":1675426260000}
[...]

api_getPriceHistoryAsDataFrame

TDAmeritradeAPI.api_getPriceHistoryAsDataFrame — Function

api_getPriceHistoryAsDataFrame(symbol::String, keys::apiKeys; kw...)

Make the TDAmeritradeAPI call to the get_price_history endpoint, and return a DataFrame

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: The ticker symbol to fetch price history for.

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

periodType::String: The type of period to show. (default in bold):

Valid values are day, month, year, or ytd (year to date)

numPeriods::Int64: The number of periods to show.

Valid periods by periodType (defaults in bold):

day: 1, 2, 3, 4, 5, 10

month: 1, 2, 3, 6

year: 1, 2, 3, 5, 10, 15, 20

ytd: 1

frequencyType::String: The type of frequency with which a new candle is formed.

Valid frequencyTypes by periodType (defaults in bold):

day: minute

month: daily, weekly

year: daily, weekly, monthly

ytd: daily, weekly

frequency::Int64: The number of the frequencyType to be included in each candle.

Valid frequencies by frequencyType (defaults in bold):

minute: 1, 5, 10, 15, 30

daily: 1

weekly: 1

monthly: 1

endDate::DateTime: End date of data to fetch. If startDate and endDate are provided, period should not be provided.

startDate::Union{DateTime, Nothing}: Start date of data to fetch.

needExtendedHoursData: true to return extended hours data, false for regular market hours only. Default is true.

Example

api_getPriceHistoryAsDataFrame("NET", apiKey)
some(5216x6 DataFrame
  Row | Datetime             Open     High     Low      Close    Volume
      | DateTime             Float64  Float64  Float64  Float64  Int64
  --------------------------------------------------------------------
    1 | 2023-02-03T12:00:00  60.8     60.8     60.5     60.5       5000
    2 | 2023-02-03T12:08:00  61.0     61.0     61.0     61.0        200
    3 | 2023-02-03T12:11:00  61.75    61.8     61.75    61.8       2347
    4 | 2023-02-03T12:14:00  61.61    61.61    61.61    61.61       162
    5 | 2023-02-03T12:17:00  61.8     61.8     61.79    61.79       228
    6 | 2023-02-03T12:26:00  61.3     61.3     61.3     61.3        159
    7 | 2023-02-03T12:31:00  61.43    61.43    61.43    61.43       100
    8 | 2023-02-03T12:34:00  61.5     61.5     61.5     61.5        102
    9 | 2023-02-03T12:35:00  61.5     61.5     61.5     61.5        400
   10 | 2023-02-03T12:38:00  61.57    61.61    61.57    61.61       206
[...]

api_getPriceHistoryAsTimeArray

TDAmeritradeAPI.api_getPriceHistoryAsTimeArray — Function

api_getPriceHistoryAsTimeArray(symbol::String, keys::apiKeys; kw...)

Make the TDAmeritradeAPI call to the get_price_history endpoint, and return a TimeSeries.TimeArray

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: The ticker symbol to fetch price history for.

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Keywords

periodType::String: The type of period to show. (default in bold):

Valid values are day, month, year, or ytd (year to date)

numPeriods::Int64: The number of periods to show.

Valid periods by periodType (defaults in bold):

day: 1, 2, 3, 4, 5, 10

month: 1, 2, 3, 6

year: 1, 2, 3, 5, 10, 15, 20

ytd: 1

frequencyType::String: The type of frequency with which a new candle is formed.

Valid frequencyTypes by periodType (defaults in bold):

day: minute

month: daily, weekly

year: daily, weekly, monthly

ytd: daily, weekly

frequency::Int64: The number of the frequencyType to be included in each candle.

Valid frequencies by frequencyType (defaults in bold):

minute: 1, 5, 10, 15, 30

daily: 1

weekly: 1

monthly: 1

endDate::DateTime: End date of data to fetch. If startDate and endDate are provided, period should not be provided.

startDate::Union{DateTime, Nothing}: Start date of data to fetch.

needExtendedHoursData: true to return extended hours data, false for regular market hours only. Default is true.

Example

api_getPriceHistoryAsTimeArray("NET", apiKey)
Option{TimeArray}(some(5216x5 TimeArray{Float64, 2, DateTime, Matrix{Float64}} 2023-02-03T12:00:00 to 2023-02-17T00:57:00
|                     | Open  | High  | Low   | Close | Volume |
----------------------------------------------------------------
| 2023-02-03T12:00:00 | 60.8  | 60.8  | 60.5  | 60.5  | 5000.0 |
| 2023-02-03T12:08:00 | 61.0  | 61.0  | 61.0  | 61.0  | 200.0  |
| 2023-02-03T12:11:00 | 61.75 | 61.8  | 61.75 | 61.8  | 2347.0 |
| 2023-02-03T12:14:00 | 61.61 | 61.61 | 61.61 | 61.61 | 162.0  |
| 2023-02-03T12:17:00 | 61.8  | 61.8  | 61.79 | 61.79 | 228.0  |
[...]

parsePriceHistoryJSONToDataFrame

TDAmeritradeAPI.parsePriceHistoryJSONToDataFrame — Function

parsePriceHistoryJSONToDataFrame(json_string::String)::ErrorTypes.Option{DataFrame}

Convert the JSON string returned by a TDAmeritradeAPI get_price_history API to a DataFrame.

Nested JSON objects will be flattened into columns in the output DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

parsePriceHistoryJSONToDataFrame(j)
some(5216x6 DataFrame
  Row | Datetime             Open     High     Low      Close    Volume
      | DateTime             Float64  Float64  Float64  Float64  Int64
  --------------------------------------------------------------------
    1 | 2023-02-03T12:00:00  60.8     60.8     60.5     60.5       5000
    2 | 2023-02-03T12:08:00  61.0     61.0     61.0     61.0        200
    3 | 2023-02-03T12:11:00  61.75    61.8     61.75    61.8       2347
    4 | 2023-02-03T12:14:00  61.61    61.61    61.61    61.61       162
    5 | 2023-02-03T12:17:00  61.8     61.8     61.79    61.79       228
    6 | 2023-02-03T12:26:00  61.3     61.3     61.3     61.3        159
    7 | 2023-02-03T12:31:00  61.43    61.43    61.43    61.43       100
    8 | 2023-02-03T12:34:00  61.5     61.5     61.5     61.5        102
    9 | 2023-02-03T12:35:00  61.5     61.5     61.5     61.5        400
   10 | 2023-02-03T12:38:00  61.57    61.61    61.57    61.61       206
[...]

parsePriceHistoryJSONToTimeArray

TDAmeritradeAPI.parsePriceHistoryJSONToTimeArray — Function

parsePriceHistoryJSONToTimeArray(json_string::String)::ErrorTypes.Option{TimeSeries.TimeArray}

Convert the JSON string returned by a TDAmeritradeAPI get_price_history API to a TimeSeries.TimeArray.

Nested JSON objects will be flattened into columns in the output TimeSeries.TimeArray.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

parsePriceHistoryJSONToTimeArray(j)
Option{TimeArray}(some(5216x5 TimeArray{Float64, 2, DateTime, Matrix{Float64}} 2023-02-03T12:00:00 to 2023-02-17T00:57:00
|                     | Open  | High  | Low   | Close | Volume |
----------------------------------------------------------------
| 2023-02-03T12:00:00 | 60.8  | 60.8  | 60.5  | 60.5  | 5000.0 |
| 2023-02-03T12:08:00 | 61.0  | 61.0  | 61.0  | 61.0  | 200.0  |
| 2023-02-03T12:11:00 | 61.75 | 61.8  | 61.75 | 61.8  | 2347.0 |
| 2023-02-03T12:14:00 | 61.61 | 61.61 | 61.61 | 61.61 | 162.0  |
| 2023-02-03T12:17:00 | 61.8  | 61.8  | 61.79 | 61.79 | 228.0  |
[...]

priceHistoryToCandleListStruct

TDAmeritradeAPI.priceHistoryToCandleListStruct — Function

priceHistoryToCandleListStruct(json_string::String)::ErrorTypes.Option{CandleList}

Convert the JSON string returned by a TDAmeritradeAPI get_price_history API call to a CandleList struct.

This is largely an internal function to allow later conversions to DataFrame with proper type conversions.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

priceHistoryToCandleListStruct(j)
some(TDAmeritradeAPI.CandleList("NET", false, TDAmeritradeAPI.Candle[TDAmeritradeAPI.Candle(1675425600000, 60.8, 60.8, 60.5, 60.5, 5000), 
TDAmeritradeAPI.Candle(1675426080000, 61.0, 61.0, 61.0, 61.0, 200),
TDAmeritradeAPI.Candle(1675426260000, 61.75, 61.8, 61.75, 61.8, 2347),
TDAmeritradeAPI.Candle(1675426440000, 61.61, 61.61, 61.61, 61.61, 162),
[...]  

priceHistoryToJSON

TDAmeritradeAPI.priceHistoryToJSON — Function

priceHistoryToJSON(cl::CandleList)::ErrorTypes.Option{String}

Convert a CandleList struct cl to a JSON object.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

The returned JSON will not be in the same format as the initial return from the TDAmeritradeAPI call.

Example

priceHistoryToJSON(s)
some("{"symbol":"NET","empty":false,"candles":[{"datetime":1675425600000,"open":60.8,"high":60.8,"low":60.5,"close":60.5,"volume":5000},
{"datetime":1675426080000,"open":61.0,"high":61.0,"low":61.0,"close":61.0,"volume":200},
[...]