Quotes

api_getQuoteAsJSON

TDAmeritradeAPI.api_getQuoteAsJSON — Function

api_getQuoteAsJSON(symbol::String, apiKeys::TDAmeritradeAPI.apiKeys)::ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quote endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: The ticker symbol to fetch a quote for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuoteAsJSON("NET", apiKey)
Result{String, String}(Ok("{"NET":{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"18915M107","assetSubType":"","symbol":"NET"
[...]

api_getQuotesAsJSON

TDAmeritradeAPI.api_getQuotesAsJSON — Function

api_getQuotesAsJSON(symbols::Vector{String}, apiKeys::TDAmeritradeAPI.apiKeys)ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quotes endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbols::Vector{String}: A Vector of ticker symbols to fetch a quotes for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuotesAsJSON(["NET", "PANW", "ZS"], apiKey)
Result{String, String}(Ok("{"NET":{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"18915M107","assetSubType":"","symbol":"NET" ...
                            "PANW":{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"697435105","assetSubType":"","symbol":"PANW" ...
[...]

api_getQuotesAsJSON(symbols::String, apiKeys::TDAmeritradeAPI.apiKeys)ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quotes endpoint, and return the raw JSON.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbols::String: A comma separated string of ticker symbols to fetch a quotes for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuotesAsJSON("NET,PANW,ZS", apiKey)
Result{String, String}(Ok("{"NET":{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"18915M107","assetSubType":"","symbol":"NET" ...
                            "PANW":{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"697435105","assetSubType":"","symbol":"PANW" ...
[...]

api_getQuoteAsDataFrame

TDAmeritradeAPI.api_getQuoteAsDataFrame — Function

api_getQuoteAsDataFrame(symbol::String, apiKeys::TDAmeritradeAPI.apiKeys)ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quote endpoint, and return a DataFrame

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbol::String: The ticker symbol to fetch a quote for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuoteAsDataFrame("NET", apiKey) 
some(1x90 DataFrame
 Row | assetType  assetMainType  cusip      assetSubType  symbol  description ...
     | String     String         String     String        String  String     
 ------------------------------------------------------------------------------------------------
   1 | EQUITY     EQUITY         18915M107                NET     Cloudflare, Inc. Class A Common
[...]                         

api_getQuotesAsDataFrame

TDAmeritradeAPI.api_getQuotesAsDataFrame — Function

api_getQuotesAsDataFrame(symbols::Vector{String}, apiKeys::TDAmeritradeAPI.apiKeys)ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quotes endpoint, and return a DataFrame

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbols::Vector{String}: A Vector of ticker symbols to fetch a quotes for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuotesAsJSON(["NET", "PANW", "ZS"], apiKey)
some(3x90 DataFrame
 Row | assetType  assetMainType  cusip      assetSubType  symbol  description ...
     | String     String         String     String        String  String      
 -------------------------------------------------------------------------------------------------
   1 | EQUITY     EQUITY         18915M107                NET     Cloudflare, Inc. Class A Common 
   2 | EQUITY     EQUITY         697435105                PANW    Palo Alto Networks, Inc. - Commo
   3 | EQUITY     EQUITY         98980G102                ZS      Zscaler, Inc. - Common Stock
[...]

api_getQuotesAsDataFrame(symbols::String, apiKeys::TDAmeritradeAPI.apiKeys)ErrorTypes.Result{String, String}

Make the TDAmeritradeAPI call to the get_quotes endpoint, and return a DataFrame

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Arguments

symbols::String: A comma separated string of ticker symbols to fetch a quotes for

apiKeys::TDAmeritradeAPI.apiKeys: the apiKeys object containing the CUST_KEY, access and refresh tokens.

Example

api_getQuotesAsJSON("NET,PANW,ZS", apiKey)
some(3x90 DataFrame
 Row | assetType  assetMainType  cusip      assetSubType  symbol  description ...
     | String     String         String     String        String  String      
 -------------------------------------------------------------------------------------------------
   1 | EQUITY     EQUITY         18915M107                NET     Cloudflare, Inc. Class A Common 
   2 | EQUITY     EQUITY         697435105                PANW    Palo Alto Networks, Inc. - Commo
   3 | EQUITY     EQUITY         98980G102                ZS      Zscaler, Inc. - Common Stock
[...]

parseQuotesJSONToDataFrame

TDAmeritradeAPI.parseQuotesJSONToDataFrame — Function

parseQuotesJSONToDataFrame(json_string::String)::ErrorTypes.Option{DataFrame}

Convert the JSON string returned by a TDAmeritradeAPI get_quote(s) API call to a DataFrame.

Nested JSON objects will be flattened into columns in the output DataFrame.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

parseQuotesJSONToDataFrame(j)
some(3x90 DataFrame
 Row | assetType  assetMainType  cusip      assetSubType  symbol  description ...
     | String     String         String     String        String  String      
 -------------------------------------------------------------------------------------------------
   1 | EQUITY     EQUITY         18915M107                NET     Cloudflare, Inc. Class A Common 
   2 | EQUITY     EQUITY         697435105                PANW    Palo Alto Networks, Inc. - Commo
   3 | EQUITY     EQUITY         98980G102                ZS      Zscaler, Inc. - Common Stock
[...]

quotesToQuoteStruct

TDAmeritradeAPI.quotesToQuoteStruct — Function

quotesToQuoteStruct(json_string::String)::ErrorTypes.Option{QuoteArray}

Convert the JSON string returned by a TDAmeritradeAPI get_quote(s) API calls to a QuoteArray struct.

This is largely an internal function to allow later conversions to DataFrame with proper type conversions.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

Example

quotesToQuoteStruct(j)
some(TDAmeritradeAPI.QuoteStruct[TDAmeritradeAPI.QuoteStruct("EQUITY", "EQUITY", "18915M107", "", "NET", "Cloudflare, Inc. Class A Common Stock", ...
TDAmeritradeAPI.QuoteStruct("EQUITY", "EQUITY", "697435105", "", "PANW", "Palo Alto Networks, Inc. - Common Stock", ...
TDAmeritradeAPI.QuoteStruct("EQUITY", "EQUITY", "98980G102", "", "ZS", "Zscaler, Inc. - Common Stock", ...
[...]  

quotesToJSON

TDAmeritradeAPI.quotesToJSON — Function

quotesToJSON(q::QuoteArray)::ErrorTypes.Option{String}

Convert a QuoteArray struct q to a JSON object.

An ErrorTypes.jl Option object will be returned that can be evaluated with ErrorTypes.@?

The returned JSON will not be in the same format as the initial return from the TDAmeritradeAPI call.

Example

quotesToJSON(s)
some("[{"assetType":"EQUITY","assetMainType":"EQUITY","cusip":"18915M107","assetSubType":"","symbol":"NET","description":"Cloudflare, Inc. Class A Common Stock"
[...]